+++ src/main/java/roomescape/application/ReservationService.java
@@ -1,8 +1,5 @@
 package roomescape.application;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
@@ -15,8 +12,6 @@ import roomescape.dto.reservation.ReservationRequest;
 
 @Service
 public class ReservationService {
-    private static final int IN_ADVANCE_RESERVATION_DAYS = 1;
-
     private final ReservationTimeService reservationTimeService;
     private final ReservationRepository reservationRepository;
     private final ThemeRepository themeRepository;
@@ -31,21 +26,12 @@ public class ReservationService {
 
     public Reservation reserve(ReservationRequest request) {
         ReservationTime time = reservationTimeService.getReservationTime(request.timeId());
-        validateReservationInAdvance(request.date(), time.getStartAt());
         if (reservationRepository.existsByReservationDateTimeAndTheme(request.date(), time.getId(),
                 request.themeId())) {
             throw new IllegalArgumentException("이미 예약된 날짜, 시간입니다.");
         }
         Reservation reservation = new Reservation(request.name(), request.date(), time, getTheme(request.themeId()));
         return reservationRepository.save(reservation);
-    }
-
-    private void validateReservationInAdvance(LocalDate date, LocalTime time) {
-        LocalDateTime reservationDateTime = LocalDateTime.of(date, time);
-        LocalDateTime baseDateTime = LocalDateTime.now().plusDays(IN_ADVANCE_RESERVATION_DAYS);
-        if (reservationDateTime.isBefore(baseDateTime)) {
-            throw new IllegalArgumentException(String.format("예약은 최소 %d일 전에 해야합니다.", IN_ADVANCE_RESERVATION_DAYS));
-        }
     }
 
     private Theme getTheme(long themeId) {
